Lombokifying the project
Had pinged Erik before when we first forked the project asking if he
would be interested in taking Lombok as a dependency since I've found in
other projects that it tends to make things more maintainable and
deliberate, and it looked like there was quite a bit of the code that
lends itself to those patterns.

Taking on that switch before working on the other feature requests I
have for the project.